fix error in settings UI on undefined accelerator
This changes the check for a `null` accelerator to catch either `null`
or `undefined`. The triple-equals operator, `=== null`, does not allow
type coercion, and will only match `null`. On the other hand the
double-equals operator, `== null`, *does* allow type coercion and as
a result considers `null` and `undefined` to be equivalent. (Although
`null` and `undefined` are isomorphic to each other, they are considered
to be distinct values which makes them distinct under strict equality
checks.)

Another option would be to check for falsy values: `if (!accelerator) {...}`
That catches `null`, `undefined`, and empty strings which you probably
want to treat the same anyway.

Fixes #3