Sort windows by user time (#23)
* Sort windows by user time

When switching monitors, we would sometimes focus a covered window (ie.
a window hidden underneath another window), which makes focus changing a
bit jarring.

This modification sorts the candidate windows by the user_time attribute
whereby a higher user_time means that the window was more recently
interactive with (and thus more likely to be in the foreground).

This is based on the isAbove function from Adjacent-Windows:

    https://github.com/klangman/Adjacent-Windows/blob/81fa84757c820df2456e3f5d145d2b8670c4777b/adjacent-windows%40klangman/5.4/extension.js#L44

This should help address issue:

    https://github.com/martinhjartmyr/gnome-shell-extension-focus-changer/issues/4

* formatting

---------

Co-authored-by: Martin Hjärtmyr <martin@hjartmyr.se>